+++ index.html
@@ -235,7 +235,7 @@
     </section>
   </template>
   <script defer src="/js/functions.js"></script>
-  <script defer src="/js/main.js"></script>
+  <script defer src="/js/main.js" type="module"></script>
 </body>
 
 </html>
